Update README.md
These quick fixes get the example code working on 0.30.5 due to changes made in gleam-lang/gleam#2047.

I added the imports, as it took a while for me (learning gleam) to figure out the errors and to add that, so hopefully it helps someone new. I don't think it'd be adverse to anyone else.